refactor for cleaner separation
