+++ shared/race-emulator/include/race-emulator/Util.h
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 #pragma once
+#include "Dim3d.h"
 #include <cassert>
 #include <charconv>
 #include <cstdint>
@@ -14,18 +15,6 @@
 #include <vector>
 
 namespace raceemulator {
-
-/// 3D integer coordinate for workgroup IDs and block dimensions.
-class Dim3d {
-public:
-  Dim3d() = delete;
-  explicit Dim3d(int x) : x(x), y(0), z(0) {}
-  Dim3d(int x, int y) : x(x), y(y), z(0) {}
-  Dim3d(int x, int y, int z) : x(x), y(y), z(z) {}
-  int x = 0;
-  int y = 0;
-  int z = 0;
-};
 
 /// Split an assembly line into whitespace/comma-separated tokens.
 std::vector<std::string_view> getPartitioned(std::string_view line);
@@ -129,26 +118,6 @@ template <typename T> T getFloatFromView(std::string_view nStr) {
                 "Requires C++20 std::from_chars for floating point parsing");
 #endif
   return value;
-}
-
-/// Iterate over active lanes of an exec mask, calling func(lane) for each.
-/// waveSize must be 32 or 64 (the two GPU wave sizes). When all lanes within
-/// the wave are active, a fast path avoids per-bit checking.
-template <typename F>
-void forEachActiveLane(uint64_t execMask, int waveSize, F func) {
-  uint64_t fullMask = (waveSize == 64) ? ~0ULL : ((1ULL << waveSize) - 1);
-
-  if ((execMask & fullMask) == fullMask) {
-    for (int lane = 0; lane < waveSize; ++lane) {
-      func(lane);
-    }
-  } else {
-    for (int lane = 0; lane < waveSize; ++lane) {
-      if ((execMask >> lane) & 1) {
-        func(lane);
-      }
-    }
-  }
 }
 
 } // namespace raceemulator
